added monitoring of battery usage, removed unnecessary execution service

+++ mobile-cloud-computing-library/src/main/java/mcc/agh/edu/pl/mobilecloudcomputinglibrary/model/KnowledgeInstance.java
@@ -3,12 +3,12 @@ package mcc.agh.edu.pl.mobilecloudcomputinglibrary.model;
 public class KnowledgeInstance {
 
     private String taskName;
-    private int batteryUsage;
+    private long batteryUsage;
     private long timeUsage;
     private boolean wifiEnabled;
     private ExecutionEnvironment env;
 
-    public KnowledgeInstance(String taskName, int batteryUsage, long timeUsage, boolean wifiEnabled, ExecutionEnvironment env){
+    public KnowledgeInstance(String taskName, long batteryUsage, long timeUsage, boolean wifiEnabled, ExecutionEnvironment env){
         this.taskName = taskName;
         this.batteryUsage = batteryUsage;
         this.timeUsage = timeUsage;
@@ -20,7 +20,7 @@ public class KnowledgeInstance {
         return taskName;
     }
 
-    public int getBatteryUsage() {
+    public long getBatteryUsage() {
         return batteryUsage;
     }
